Re-named MySQL tests so they'd actually run in order.
