+++ Makefile
@@ -1,4 +1,4 @@
-TESTS = test/mssql/*.js
+TESTS = test/mysql/*.js test/mssql/*.js
 test:
 	mocha --timeout 5000 --reporter spec $(TESTS)
