+++ test/mysql/00-AA-tests-general.js
@@ -1,0 +1,9 @@
+const should = require('chai').should();
+
+describe('QueryBuilder', () => {
+	it('actually exists and can be initialized', () => {
+		const QueryBuilder = require('../../drivers/mysql/query_builder.js');
+		const qb = new QueryBuilder();
+		qb.should.be.instanceOf(QueryBuilder);
+	});
+});
