+++ test/mysql/00-tests-adapter.js
