+++ test/mysql/02-tests-compilation_methods.js
