+++ test/mysql/03-tests-delete.js
