+++ test/mysql/03-tests-empty_table.js
