+++ test/mysql/03-tests-insert.js
