+++ test/mysql/03-tests-insert_batch.js
