+++ test/mysql/03-tests-truncate.js
