+++ test/mysql/03-tests-update_batch.js
